- minor optimizations to PFB compiler
- removed second baf_list input
- added additional options for testing a tuning
    - # of batches to run concurrently
    - # of samples to load per batch
- updated problematic regions to match CNV-Annotation
- Added autosome only mode
- General formatting and name fixes

+++ modules/CNV_calling_per_sampleID/resources/bin/cnv_calling.sh
@@ -66,7 +66,7 @@ log_step() {
 define_usage() {
   echo "Usage: $0 [--sample_id ID] [--BAF_LRR_Probes FILE] [--sexfile FILE] \\
                 [--pfb FILE] [--hmm FILE] [--gcmodel FILE] [--chr CHR] \\
-                [--levels FILE] [--config FILE] [--gcdir DIR]"
+                [--levels FILE] [--config FILE] [--gcdir DIR] [--autosome_only]"
   exit 1
 }
 
@@ -81,6 +81,7 @@ while [[ $# -gt 0 ]]; do
     --pfb) pfb_file="$2"; shift 2;;
     --hmm) hmm_file="$2"; shift 2;;
     --gcmodel) gcmodel_file="$2"; shift 2;;
+    --autosome_only) autosome_only=1; shift ;;
     # ================================
     # QuantiSNP-specific input parameters
     --chr) chr="$2"; shift 2;;
@@ -97,24 +98,27 @@ log_step "STEP 0: Starting CNV calling for sample ${sample_id}"
 # -----------------------------------------------------------------------------
 # Step 1: Extract Gender
 # -----------------------------------------------------------------------------
+if [[ "$autosome_only" -eq 0 ]]; then 
+    log_step "STEP 1: Extracting gender from sex file..."
 
-log_step "STEP 1: Extracting gender from sex file..."
+    sexfile_temp=$(mktemp)
+    zgrep -P "${sample_id}\t" ${sex_file} | cut -f1,2 > ${sexfile_temp}
+    gender=$(cut -f2 ${sexfile_temp} | tr '[:upper:]' '[:lower:]')
+    if [[ -z "$gender" ]]; then
+        log_step "ERROR: No gender information found for sample ID ${sample_id}."
+        exit 1
+    fi
 
-sexfile_temp=$(mktemp)
-zgrep -P "${sample_id}\t" ${sex_file} | cut -f1,2 > ${sexfile_temp}
-gender=$(cut -f2 ${sexfile_temp} | tr '[:upper:]' '[:lower:]')
-if [[ -z "$gender" ]]; then
-    echo "ERROR: No gender information found for sample ID ${sample_id}."
-    exit 1
+    log_step "INFO: Gender detected as '${gender}'"
+    if [[ "$gender" != "male" && "$gender" != "female" ]]; then
+        log_step "ERROR: Invalid gender '$gender' for sample ID ${sample_id}. Expected 'male' or 'female'."
+        exit 1
+    fi
 fi
 
-log_step "INFO: Gender detected as '${gender}'"
-if [[ "$gender" != "male" && "$gender" != "female" ]]; then
-    log_step "ERROR: Invalid gender '$gender' for sample ID ${sample_id}. Expected 'male' or 'female'."
-    exit 1
+if [[ "$autosome_only" -eq 0 ]]; then
+    gender="unknown"
 fi
-
-
 # -----------------------------------------------------------------------------
 # Step 2: PennCNV (Autosomes)
 # -----------------------------------------------------------------------------
@@ -138,20 +142,20 @@ timedev -v penncnv --test \
 # -----------------------------------------------------------------------------
 # Step 3: PennCNV (ChrX)
 # -----------------------------------------------------------------------------
-
-log_step "STEP 3: Running PennCNV on ChrX..."
-timedev -v penncnv --test \
-    --conf \
-    --minsnp 1 \
-    --pfbfile ${pfb_file} \
-    --hmmfile ${hmm_file} \
-    --logfile ${sample_id}.penncnv.chrx.qc \
-    --output ${sample_id}.penncnv.chrx.out \
-    --gcmodelfile ${gcmodel_file} \
-    --sexfile ${sexfile_temp} \
-    --chrx \
-    ${BAF_LRR_Probes} &> ${sample_id}.penncnv.chrx.log
-
+if [[ "$autosome_only" -eq 0 ]]; then  
+    log_step "STEP 3: Running PennCNV on ChrX..."
+    timedev -v penncnv --test \
+        --conf \
+        --minsnp 1 \
+        --pfbfile ${pfb_file} \
+        --hmmfile ${hmm_file} \
+        --logfile ${sample_id}.penncnv.chrx.qc \
+        --output ${sample_id}.penncnv.chrx.out \
+        --gcmodelfile ${gcmodel_file} \
+        --sexfile ${sexfile_temp} \
+        --chrx \
+        ${BAF_LRR_Probes} &> ${sample_id}.penncnv.chrx.log
+fi
 
 # -----------------------------------------------------------------------------
 # Step 4: QuantiSNP
@@ -159,18 +163,30 @@ timedev -v penncnv --test \
 
 log_step "STEP 4: Running QuantiSNP..."
 export MCR_CACHE_ROOT=$(mktemp -d mcr_cache_XXXXXX)
+if [[ "$autosome_only" -eq 0 ]]; then
+    timedev -v quantisnp --chr ${chr} \
+        --outdir . \
+        --sampleid ${sample_id} \
+        --gender ${gender} \
+        --config ${config} \
+        --levels ${levels} \
+        --gcdir ${gcdir} \
+        --input-files ${BAF_LRR_Probes} \
+        --doXcorrect --verbose &> ${sample_id}.quantisnp.log
 
-timedev -v quantisnp --chr ${chr} \
-    --outdir . \
-    --sampleid ${sample_id} \
-    --gender ${gender} \
-    --config ${config} \
-    --levels ${levels} \
-    --gcdir ${gcdir} \
-    --input-files ${BAF_LRR_Probes} \
-    --doXcorrect --verbose &> ${sample_id}.quantisnp.log
+    rm -rf "$MCR_CACHE_ROOT"
+else
+    timedev -v quantisnp --chr ${chr} \
+        --outdir . \
+        --sampleid ${sample_id} \
+        --config ${config} \
+        --levels ${levels} \
+        --gcdir ${gcdir} \
+        --input-files ${BAF_LRR_Probes} 
 
-rm -rf "$MCR_CACHE_ROOT"
+    rm -rf "$MCR_CACHE_ROOT"
+fi
+
 
 # -----------------------------------------------------------------------------
 # Step 5: Renaming files
@@ -182,12 +198,16 @@ mv ${sample_id}.loh ${sample_id}.quantisnp.loh
 mv ${sample_id}.qc ${sample_id}.quantisnp.qc
 
 log_step "STEP 6: Merging PennCNV results..."
-cat ${sample_id}.penncnv.chrx.out ${sample_id}.penncnv.out > ${sample_id}.penncnv.cnv
+if [[ "$autosome_only" -eq 0 ]]; then 
+    cat ${sample_id}.penncnv.chrx.out ${sample_id}.penncnv.out > ${sample_id}.penncnv.cnv
 
-rm ${sample_id}.penncnv.chrx.out ${sample_id}.penncnv.out
+    rm ${sample_id}.penncnv.chrx.out ${sample_id}.penncnv.out
 
-rm ${sexfile_temp}
+    rm ${sexfile_temp}
+else
+    mv ${sample_id}.penncnv.out ${sample_id}.penncnv.cnv
 
+fi
 log_step "STEP 7: Formatting QuantiSNP and PennCNV CNV files..."
 format_quantisnp_cnv.sh "${sample_id}.quantisnp.cnv" "${sample_id}.quantisnp.cnv.tsv"
 format_penncnv_cnv.sh "${sample_id}.penncnv.cnv" "${sample_id}.penncnv.cnv.tsv"
